# beacon-server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "MeshCore network observation backend. Ingests LoRa packets from MQTT brokers, stores in PostgreSQL, and streams live events via WebSocket.",
        "title": "MeshCore Beacon API",
        "termsOfService": "https://github.com/MeshCore-Beacon/beacon-server",
        "contact": {
            "name": "MeshCore Beacon",
            "url": "https://github.com/MeshCore-Beacon/beacon-server"
        },
        "license": {
            "name": "AGPL-3-or-later"
        },
        "version": "1.6.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/routes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Routes"
                ],
                "summary": "List known routes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by IATA code (legacy singular form)",
                        "name": "iata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by IATA codes, comma-separated",
                        "name": "iatas",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by exact hop count",
                        "name": "hopCount",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Epoch ms timestamp of last item for pagination",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque keyset cursor returned as nextPageToken",
                        "name": "pageToken",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by iata, hops, observations, first_seen or last_seen",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction: asc or desc",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_KnownRoute"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/routes/cross": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Routes"
                ],
                "summary": "Search for routes that cross IATA boundaries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source node hash prefix (hex)",
                        "name": "fromHash",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Source IATA code",
                        "name": "fromIata",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Destination node hash prefix (hex)",
                        "name": "toHash",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Destination IATA code",
                        "name": "toIata",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.CrossIATARoute"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/routes/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Routes"
                ],
                "summary": "Search known routes by source and destination hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IATA code to search within",
                        "name": "iata",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Source node hash prefix (hex)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Destination node hash prefix (hex)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.KnownRoute"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.CrossIATAHop": {
            "type": "object",
            "properties": {
                "fromIata": {
                    "type": "string"
                },
                "fromNode": {
                    "description": "last node in source IATA",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode"
                        }
                    ]
                },
                "lastSeen": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "toIata": {
                    "type": "string"
                },
                "toNode": {
                    "description": "first node in target IATA",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode"
                        }
                    ]
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.CrossIATARoute": {
            "type": "object",
            "properties": {
                "crossHop": {
                    "description": "the boundary hop",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.CrossIATAHop"
                        }
                    ]
                },
                "sourceSegment": {
                    "description": "route segment in source IATA",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RouteHop"
                    }
                },
                "targetSegment": {
                    "description": "route segment in target IATA",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RouteHop"
                    }
                },
                "totalHops": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.KnownRoute": {
            "type": "object",
            "properties": {
                "firstSeen": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "hopCount": {
                    "type": "integer"
                },
                "hops": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RouteHop"
                    }
                },
                "iata": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastSeen": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "observationCount": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_KnownRoute": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.KnownRoute"
                    }
                },
                "nextCursor": {
                    "type": "integer"
                },
                "nextPageToken": {
                    "type": "string"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "publicKey": {
                    "description": "hex-encoded prefix used for resolution",
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.RouteHop": {
            "type": "object",
            "properties": {
                "hashBytes": {
                    "description": "hex-encoded hash prefix",
                    "type": "string"
                },
                "node": {
                    "description": "populated when node details are available",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode"
                        }
                    ]
                },
                "nodeId": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g. \"not_found\", \"bad_request\"",
                    "type": "string"
                },
                "message": {
                    "description": "e.g. \"channel not found\"",
                    "type": "string"
                }
            }
        },
...
    },
...
}